Refactor: Move map line parsing functions to `load_data_utils3.c`
